deep copy and fix.
