+++ helper/deepCopy.js
@@ -1,0 +1,18 @@
+export function deepCopy(inObject) {
+	if (typeof inObject !== 'object' || inObject === null) {
+		return inObject; // Return the value if inObject is not an object
+	}
+
+	// Create an array or object to hold the values
+	const outObject = Array.isArray(inObject) ? [] : {};
+
+	let value;
+	for (const key in inObject) {
+		value = inObject[key];
+
+		// Recursively (deep) copy for nested objects, including arrays
+		outObject[key] = deepCopy(value);
+	}
+
+	return outObject;
+}
